Add folder AuthForm to components and move AuthForm.jsx there
